Fix warnings: invalid integer narrowing.

+++ InputSettings.h
@@ -156,8 +156,8 @@ class CInputSettings : CSettings
 			const char*	pszDescription;	// Description of key.
 			const char*	pszSaveName;		// Name for INI.
 			U8		u8DefaultKey;		// Default rspScanKeys val (RSP_SK_*).
-			short	sDefMouseButtons;	// Default rspGetMouse psButtons mask (MouseButtons).
-			short	sDefJoyButtons;	// Default rspGetJoyState buttons mask (JoyButtons).
+			int		sDefMouseButtons;	// Default rspGetMouse psButtons mask (MouseButtons).
+			int		sDefJoyButtons;	// Default rspGetJoyState buttons mask (JoyButtons).
 			} InputInfo;
